Changed the Projection API so that the projected points are also now represented by Point2D objects.  All the other classes changed as a result.
git-svn-id: svn+ssh://src.dsl.bbn.com/svn/openmap/trunk@1601 f68c23cc-7f87-4fbc-acfa-d9760b90acc1

+++ src/openmap/com/bbn/openmap/proj/LLXYView.java
@@ -14,8 +14,8 @@
 // 
 // $Source: /cvs/distapps/openmap/src/openmap/com/bbn/openmap/proj/LLXYView.java,v $
 // $RCSfile: LLXYView.java,v $
-// $Revision: 1.5 $
-// $Date: 2005/12/09 21:09:02 $
+// $Revision: 1.6 $
+// $Date: 2006/02/16 16:22:46 $
 // $Author: dietrick $
 // 
 // **********************************************************************
@@ -290,7 +290,7 @@ public class LLXYView extends LLXY {
             this.computeParameters();
         }
 
-        public void setAllParams(int hPixelsPerMeter, double hPlanetRadius,
+        public void setAllParams(double hPixelsPerMeter, double hPlanetRadius,
                                  double hPlanetPixelRadius,
                                  double hPlanetPixelCircumference,
                                  double hMinscale, double hMaxscale,
